requirements.txt
Flask==2.0.3
google-genai==0.1.0
python-dotenv==0.19.2
